chatbot/requirements.txt
openai==0.14.0
psycopg2==2.9.3
scipy==1.8.0
py-cord==2.0.0
python-dotenv==0.19.2
sentry-sdk==1.9.3
amplitude-analytics==0.4.1
aiohttp==3.8.1